feat: lazy load download part
